Use Lychee directly instead of going through Baler
Baler seems to me to only be able to process code that has been committed to the Git repository. In the case of nmdc-schema, I want to test the generated Markdown documentation, which is not committed to the Git repository. Lychee supports that. The Lychee (action) docs also include an example of creating a GitHub Issue listing the broken links it finds, which was one of the selling points of Baler in the first place.