Update basic_slots.yaml
Normalize the mentions of other elements in DataCategoryEnum permissible value descriptions so that they match the mentioned elements name exactly 

Principle: when someone is reading the description of a schema element, they should be able to copy the mentions of elements and search for them verbatim, even with case sensitivity on. 

There are still some other violations of this policy in this YAML file and probably thought the other modules too.